US - 14 - backend (refactoring, modifications Coach Feedback) - frontend (error fixing)
