+++ backend/src/main/java/de/neuefische/capstone/pia/backend/controller/CalendarController.java
@@ -1,0 +1,26 @@
+package de.neuefische.capstone.pia.backend.controller;
+
+import de.neuefische.capstone.pia.backend.model.Activity;
+import de.neuefische.capstone.pia.backend.service.CalendarService;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/calendar")
+public class CalendarController {
+
+    private final CalendarService calendarService;
+
+    public CalendarController(CalendarService calendarService) {
+        this.calendarService = calendarService;
+    }
+
+    @GetMapping
+    public List<Activity> getActivitiesByMonth(
+            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate month) {
+        return calendarService.getActivitiesByMonth(month);
+    }
+}
