+++ backend/src/main/java/de/neuefische/capstone/pia/backend/controller/HobbyController.java
@@ -5,37 +5,25 @@ import de.neuefische.capstone.pia.backend.model.ActivityWithoutID;
 import de.neuefische.capstone.pia.backend.model.Hobby;
 import de.neuefische.capstone.pia.backend.model.HobbyAddModel;
 import de.neuefische.capstone.pia.backend.service.ActivityService;
-import de.neuefische.capstone.pia.backend.service.CalendarService;
 import de.neuefische.capstone.pia.backend.service.HobbyService;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/api/hobbies")
 public class HobbyController {
     private final HobbyService hobbyService;
     private final ActivityService activityService;
-    private final CalendarService calendarService;
 
-    public HobbyController(HobbyService hobbyService, ActivityService activityService, CalendarService calendarService) {
+    public HobbyController(HobbyService hobbyService, ActivityService activityService) {
         this.hobbyService = hobbyService;
         this.activityService = activityService;
-        this.calendarService = calendarService;
     }
     @GetMapping
     public List<Hobby> getHobbies() {
         return hobbyService.getHobbies();
-    }
-    @GetMapping("/calendar")
-    public List<Activity> getActivitiesByMonth(
-            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate month) {
-        return calendarService.getActivitiesByMonth(month);
     }
 
     @PostMapping
@@ -44,8 +32,8 @@ public class HobbyController {
     }
 
     @PutMapping("/{hobbyId}")
-    public Hobby updateHobby(@PathVariable String hobbyId, @RequestBody HobbyAddModel updatedHobby, @RequestParam(required = false) String color) {
-        return hobbyService.updateHobby(hobbyId, updatedHobby, color);
+    public Hobby updateHobby(@PathVariable String hobbyId, @RequestBody HobbyAddModel updatedHobby) {
+        return hobbyService.updateHobby(hobbyId, updatedHobby);
     }
 
     @PutMapping("/{hobbyId}/color")
@@ -66,7 +54,7 @@ public class HobbyController {
 
     @GetMapping("/hobby/{hobbyId}/activities")
     public List<Activity> getHobbyByIdListActivities(@PathVariable String hobbyId) {
-        return activityService.getActivitiesByHobbyId(hobbyId);
+        return activityService.getActivitiesForHobby(hobbyId);
     }
 
     @PostMapping("/hobby/{hobbyId}/activities")
@@ -84,25 +72,4 @@ public class HobbyController {
     public void deleteActivity(@PathVariable String hobbyId, @PathVariable String activityId) {
         activityService.deleteActivity(hobbyId, activityId);
     }
-
-    @GetMapping("/statistics/activity-counts")
-    public Map<String, Long> getActivityCounts() {
-        return activityService.getActivityCounts();
-    }
-
-    @GetMapping("/statistics/most-added-activity")
-    public ResponseEntity<String> getMostAddedActivity() {
-        String mostAddedActivityName = activityService.getMostAddedActivityName();
-
-        if (mostAddedActivityName != null) {
-            return ResponseEntity.ok(mostAddedActivityName);
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-    @GetMapping("/statistics/activity-days")
-    public Map<String, Long> getActivityDays() {
-        return activityService.getActivityDays();
-    }
-
 }
