+++ backend/src/main/java/de/neuefische/capstone/pia/backend/service/CalendarService.java
@@ -5,7 +5,6 @@ import de.neuefische.capstone.pia.backend.model.Hobby;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -18,28 +17,28 @@ public class CalendarService {
     }
 
     public List<Activity> getActivitiesByMonth(LocalDate month) {
-        List<Activity> activitiesWithColor = new ArrayList<>();
         List<Hobby> hobbies = hobbyService.getHobbies();
 
-        for (Hobby hobby : hobbies) {
-            List<Activity> hobbyActivities = hobby.getActivities();
-            if (hobbyActivities != null) {
-                for (Activity activity : hobbyActivities) {
-                    LocalDate activityDate = activity.getActivityDate();
-                    if (activityDate.getMonth() == month.getMonth() && activityDate.getYear() == month.getYear()) {
-                        Activity activityWithColor = new Activity();
-                        activityWithColor.setActivityId(activity.getActivityId());
-                        activityWithColor.setActivityDate(activityDate);
-                        activityWithColor.setColor(hobby.getColor());
-                        activityWithColor.setHobbyId(hobby.getHobbyId());
-                        activityWithColor.setRating(activity.getRating());
-                        activityWithColor.setName(activity.getName());
-                        activitiesWithColor.add(activityWithColor);
-                    }
-                }
-            }
-        }
+        return hobbies.stream()
+                .flatMap(hobby -> hobby.getActivities().stream()
+                        .filter(activity -> isActivityInMonth(activity, month))
+                        .map(activity -> createActivityWithColor(activity, hobby.getColor())))
+                .toList();
+    }
 
-        return activitiesWithColor;
+    private boolean isActivityInMonth(Activity activity, LocalDate month) {
+        LocalDate activityDate = activity.getActivityDate();
+        return activityDate.getMonth() == month.getMonth() && activityDate.getYear() == month.getYear();
+    }
+
+    private Activity createActivityWithColor(Activity activity, String color) {
+        Activity activityWithColor = new Activity();
+        activityWithColor.setActivityId(activity.getActivityId());
+        activityWithColor.setActivityDate(activity.getActivityDate());
+        activityWithColor.setColor(color);
+        activityWithColor.setHobbyId(activity.getHobbyId());
+        activityWithColor.setRating(activity.getRating());
+        activityWithColor.setName(activity.getName());
+        return activityWithColor;
     }
 }
