+++ backend/src/main/java/de/neuefische/capstone/pia/backend/service/HobbyService.java
@@ -1,6 +1,7 @@
 package de.neuefische.capstone.pia.backend.service;
 
 import de.neuefische.capstone.pia.backend.exceptions.NoSuchHobbyException;
+import de.neuefische.capstone.pia.backend.exceptions.UnauthorizedAccessException;
 import de.neuefische.capstone.pia.backend.model.*;
 import de.neuefische.capstone.pia.backend.repo.HobbyRepo;
 import de.neuefische.capstone.pia.backend.security.MongoUser;
@@ -8,6 +9,7 @@ import de.neuefische.capstone.pia.backend.security.MongoUserService;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import java.util.*;
+
 
 
 @Service
@@ -39,29 +41,37 @@ public class HobbyService {
         return hobbyRepo.insert(hobby);
     }
 
-    public Hobby updateHobby(String id, HobbyAddModel hobbyNoID, String newColor) {
+    public Hobby getHobbyWithUserPermission(String id) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = mongoUserService.findUserByUsername(username);
+
         Hobby hobby = hobbyRepo.findById(id)
                 .orElseThrow(() -> new NoSuchHobbyException(id));
-        hobby.setName(hobbyNoID.getName());
-        if (newColor != null) {
-            hobby.setColor(newColor);
+
+        if (!hobby.getAuthorId().equals(user.userId())) {
+            throw new UnauthorizedAccessException();
         }
+        return hobby;
+    }
+
+    public Hobby updateHobby(String id, HobbyAddModel hobbyNoID) {
+        Hobby hobby = getHobbyWithUserPermission(id);
+        hobby.setName(hobbyNoID.getName());
+        return hobbyRepo.save(hobby);
+    }
+
+    public Hobby updateHobbyColor(String id, String color) {
+        Hobby hobby = getHobbyWithUserPermission(id);
+        hobby.setColor(color);
         return hobbyRepo.save(hobby);
     }
 
     public void deleteHobby(String id) {
+        getHobbyWithUserPermission(id);
         hobbyRepo.deleteById(id);
     }
 
     public Hobby getHobbyById(String hobbyId) {
         return hobbyRepo.findById(hobbyId).orElseThrow(() -> new NoSuchHobbyException(hobbyId));
     }
-
-    public Hobby updateHobbyColor(String id, String color) {
-        Hobby hobby = hobbyRepo.findById(id)
-                .orElseThrow(() -> new NoSuchHobbyException(id));
-        hobby.setColor(color);
-        return hobbyRepo.save(hobby);
-    }
-
 }
